refactor : 건물별 전기 코드 분리
