+++ src/pages/BuildingElectricity/util.ts
@@ -1,3 +1,5 @@
+import { chartInfoType } from '../../type/Types';
+
 const dropdownInfoCreater = (
   height: string,
   xCoordinate: string,
@@ -18,4 +20,24 @@ const dropdownInfoCreater = (
   };
 };
 
-export { dropdownInfoCreater };
+const createChartCategoryArray = (
+  chart: chartInfoType[],
+  monthCategory: any
+) => {
+  const yearData = chart.map((item: chartInfoType) => item.year);
+  const yearLabel = yearData.map((year: any) => year.toString());
+  const yearTotalWaste = chart.map((item: chartInfoType) =>
+    item.usages.reduce((acc: any, cur: any) => {
+      return acc + cur.data;
+    }, 0)
+  );
+  const monthLabel = monthCategory;
+  const matchChartCategory: any = {
+    '월별 전기 사용량': ['2023', yearData, monthLabel],
+    '연별 전기 사용량': [null, null, yearLabel, yearTotalWaste],
+    '동월 전기 사용량': ['12월', monthCategory, yearLabel],
+  };
+  return matchChartCategory;
+};
+
+export { dropdownInfoCreater, createChartCategoryArray };
